+++ app/src/main/java/com/shypolarbear/android/di/ApiModule.kt
@@ -1,19 +1,19 @@
 package com.shypolarbear.android.di
 
 import com.shypolarbear.data.api.ExampleApi
-import com.shypolarbear.data.api.ImageApi
+import com.shypolarbear.data.api.image.ImageApi
 import com.shypolarbear.data.api.InfoApi
 import com.shypolarbear.data.api.JoinApi
 import com.shypolarbear.data.api.LoginApi
 import com.shypolarbear.data.api.TokenApi
 import com.shypolarbear.data.api.feed.FeedApi
+import com.shypolarbear.data.api.image.ImageUploadApi
 import com.shypolarbear.data.api.quiz.QuizApi
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import retrofit2.Retrofit
-import retrofit2.create
 import javax.inject.Singleton
 
 @Module
@@ -66,7 +66,13 @@ object ApiModule {
 
     @Singleton
     @Provides
-    fun provideImageApi(@AuthRetrofit retrofit: Retrofit): ImageApi{
+    fun provideImageApi(@AuthRetrofit retrofit: Retrofit): ImageApi {
         return retrofit.create(ImageApi::class.java)
+    }
+
+    @Singleton
+    @Provides
+    fun provideImageUploadApi(@NormalRetrofit retrofit: Retrofit): ImageUploadApi {
+        return retrofit.create(ImageUploadApi::class.java)
     }
 }
