+++ domain/src/main/java/com/shypolarbear/domain/repository/image/ImageUploadRepo.kt
@@ -1,0 +1,8 @@
+package com.shypolarbear.domain.repository.image
+
+import com.shypolarbear.domain.model.image.ImageUploadRequest
+import com.shypolarbear.domain.model.image.ImageUploadResponse
+
+interface ImageUploadRepo {
+    suspend fun imageUploadRequest(imageUploadRequest: ImageUploadRequest): Result<ImageUploadResponse>
+}
