+++ domain/src/main/java/com/shypolarbear/domain/usecase/image/ImageUploadUseCase.kt
@@ -2,10 +2,10 @@ package com.shypolarbear.domain.usecase.image
 
 import com.shypolarbear.domain.model.image.ImageUploadRequest
 import com.shypolarbear.domain.model.image.ImageUploadResponse
-import com.shypolarbear.domain.repository.ImageRepo
+import com.shypolarbear.domain.repository.image.ImageUploadRepo
 
 class ImageUploadUseCase(
-    private val repo: ImageRepo
+    private val repo: ImageUploadRepo
 ) {
     suspend operator fun invoke(imageUploadRequest: ImageUploadRequest): Result<ImageUploadResponse>{
         return repo.imageUploadRequest(imageUploadRequest)
